Refactor main.js: playback and subtitles controllers
